database works

+++ api/database/seed.js
@@ -4,9 +4,11 @@ const seedDB = async () => {
     try {
     await client.connect()
     console.log('awaiting seed...')
-    await client.db('revision_app').collection('users').drop()
+    if(client.db('revision_app')){
+        await client.db('revision_app').collection('users').drop()
+    }
     await client.db('revision_app').collection('users').insertOne(
-        { name: "Louis Theroux", email: "example@gmail.com", password: "pass" }
+        { name: "Nicolas Sanschagrin", email: "nsanschagrin7@gmail.com", password: "pass" }
     )
     console.log("DB seeded !")
     await client.close()
